TCVP-3173 fix and simplify look ups
