+++ src/backend/TrafficCourts/Citizen.Service/Services/Lookups/StatuteLookupService.cs
@@ -6,16 +6,16 @@ namespace TrafficCourts.Citizen.Service.Services.Lookups;
 
 public class StatuteLookupService : IStatuteLookupService
 {
-    private readonly ICachedLookupService<Statute> _repository;
+    private readonly IStatuteRepository _repository;
 
-    public StatuteLookupService(ICachedLookupService<Statute> repository)
+    public StatuteLookupService(IStatuteRepository repository)
     {
         _repository = repository ?? throw new ArgumentNullException(nameof(repository));
     }
 
     public async Task<TrafficCourts.Domain.Models.Statute?> GetBySectionAsync(string section, CancellationToken cancellationToken)
     {
-        var items = await GetAsync(cancellationToken);
+        var items = await _repository.GetListAsync(cancellationToken);
 
         var buffer = new StringBuilder();
         var models = items.Select(_ => _.ToDomainModel(buffer)).ToList();
@@ -25,7 +25,7 @@ public class StatuteLookupService : IStatuteLookupService
 
     public async Task<TrafficCourts.Domain.Models.Statute?> GetByIdAsync(int statuteId, CancellationToken cancellationToken)
     {
-        var items = await GetAsync(cancellationToken);
+        var items = await _repository.GetListAsync(cancellationToken);
 
         var index = items.BinarySearch(
             new Statute { stat_id = statuteId },
@@ -41,19 +41,12 @@ public class StatuteLookupService : IStatuteLookupService
 
     public async Task<IList<TrafficCourts.Domain.Models.Statute>> GetListAsync(CancellationToken cancellationToken)
     {
-        var items = await GetAsync(cancellationToken);
+        var items = await _repository.GetListAsync(cancellationToken);
 
         var buffer = new StringBuilder();
 
         var models = items.Select(x => x.ToDomainModel(buffer)).ToList();
 
         return models;
-    }
-
-    private async Task<List<Statute>> GetAsync(CancellationToken cancellationToken)
-    {
-        var key = Caching.Cache.Api.Statutes(2);
-        var items = await _repository.GetListAsync(key, TimeSpan.FromHours(2), cancellationToken);
-        return items;
     }
 }
