--- src/backend/TrafficCourts/Staff.Service/Features/Lookups/Cities/Handler.cs
@@ -1,6 +1,0 @@
-﻿namespace TrafficCourts.Staff.Service.Features.Lookups.Cities
-{
-    public class Handler
-    {
-    }
-}
