--- src/backend/TrafficCourts/Staff.Service/Features/Lookups/Cities/Request.cs
@@ -1,7 +1,0 @@
-﻿using MediatR;
-
-namespace TrafficCourts.Staff.Service.Features.Lookups.Cities;
-
-public class Request : IRequest<Response>
-{
-}
