--- src/backend/TrafficCourts/Staff.Service/Features/Lookups/Cities/Response.cs
@@ -1,16 +1,0 @@
-﻿
-namespace TrafficCourts.Staff.Service.Features.Lookups.Cities;
-
-public class City
-{
-}
-
-public record Response
-{
-    public Response(IList<City> items)
-    {
-        Items = items;
-    }
-
-    public IList<City> Items { get; }
-}
