+++ src/backend/TrafficCourts/Staff.Service/Features/Lookups/DisputeCaseFileStatusTypes/Handler.cs
@@ -1,37 +1,19 @@
 ﻿using MediatR;
 using TrafficCourts.Domain.Models;
 using TrafficCourts.OrdsDataService.Tco;
-using ZiggyCreatures.Caching.Fusion;
 
 namespace TrafficCourts.Staff.Service.Features.Lookups.DisputeCaseFileStatusTypes;
 
 public class Handler : IRequestHandler<Request, Response>
 {
     private readonly IDisputeStatusTypeRepository _repository;
-    private readonly IFusionCache _cache;
-    private readonly ILogger<Handler> _logger;
 
-    public Handler(IDisputeStatusTypeRepository repository, IFusionCache cache, ILogger<Handler> logger)
+    public Handler(IDisputeStatusTypeRepository repository)
     {
         _repository = repository ?? throw new ArgumentNullException(nameof(repository));
-        _cache = cache ?? throw new ArgumentNullException(nameof(cache));
-        _logger = logger ?? throw new ArgumentNullException(nameof(logger));
     }
 
     public async Task<Response> Handle(Request request, CancellationToken cancellationToken)
-    {
-        var key = Caching.Cache.Api.DisputeCaseFileStatusTypes(2);
-
-        var items = await _cache.GetOrSetAsync<List<DisputeCaseFileStatus>> (
-            key,
-            GetItems,
-            options => options.SetDuration(TimeSpan.FromMinutes(60)),
-            token: cancellationToken);
-
-        return new Response(items);
-    }
-
-    private async Task<List<DisputeCaseFileStatus>> GetItems(CancellationToken cancellationToken)
     {
         var items = await _repository.GetListAsync(cancellationToken);
 
@@ -41,6 +23,6 @@ public class Handler : IRequestHandler<Request, Response>
             Description = item.dispute_status_type_dsc
         }).ToList();
 
-        return models;
+        return new Response(models);
     }
 }
