+++ src/backend/TrafficCourts/Staff.Service/Features/Lookups/Languages/Handler.cs
@@ -1,44 +1,21 @@
 ﻿using MediatR;
 using TrafficCourts.OrdsDataService.Justin;
-using ZiggyCreatures.Caching.Fusion;
+using TrafficCourts.Staff.Service.Services;
 
 namespace TrafficCourts.Staff.Service.Features.Lookups.Languages;
 
 public class Handler : IRequestHandler<Request, Response>
 {
-    private readonly ILanguageRepository _repository;
-    private readonly IFusionCache _cache;
-    private readonly ILogger<Handler> _handler;
+    private readonly ILanguageLookupService _service;
 
-    public Handler(ILanguageRepository repository, IFusionCache cache, ILogger<Handler> handler)
+    public Handler(ILanguageLookupService service)
     {
-        _repository = repository ?? throw new ArgumentNullException(nameof(repository));
-        _cache = cache ?? throw new ArgumentNullException(nameof(cache));
-        _handler = handler ?? throw new ArgumentNullException(nameof(handler));
+        _service = service ?? throw new ArgumentNullException(nameof(service));
     }
+
     public async Task<Response> Handle(Request request, CancellationToken cancellationToken)
     {
-        var key = Caching.Cache.Api.Languages(2);
-
-        var items = await _cache.GetOrSetAsync<List<Domain.Models.Language>>(
-            key,
-            ct => GetItems(ct),
-            options => options.SetDuration(TimeSpan.FromMinutes(10)),
-            token: cancellationToken);
-
+        var items = await _service.GetListAsync(cancellationToken);
         return new Response(items);
-    }
-
-    private async Task<List<Domain.Models.Language>> GetItems(CancellationToken cancellationToken)
-    {
-        var items = await _repository.GetListAsync(cancellationToken);
-
-        var models = items.Select(_ => new Domain.Models.Language
-        (
-            _.cdln_language_cd,
-            _.cdln_language_dsc
-        )).ToList();
-
-        return models;
     }
 }
