+++ src/backend/TrafficCourts/Staff.Service/Services/ILanguageLookupService.cs
@@ -1,0 +1,6 @@
+﻿namespace TrafficCourts.Staff.Service.Services;
+
+public interface ILanguageLookupService
+{
+    Task<List<Domain.Models.Language>> GetListAsync(CancellationToken cancellationToken);
+}
