+++ src/backend/TrafficCourts/Staff.Service/Services/IStatuteLookupService.cs
@@ -1,0 +1,6 @@
+﻿namespace TrafficCourts.Staff.Service.Services;
+
+public interface IStatuteLookupService
+{
+    Task<Domain.Models.Statute?> GetByIdAsync(int statuteId, CancellationToken cancellationToken);
+}
